# docs/api/openapi.v1.yaml
components:
  schemas:
    AccessibilityProfile:
      additionalProperties: false
      description: 'Per-learner accessibility preferences (Phase 5, ADR-029).


        Setting `co_alternative="lexical_alt"` reroutes any CO drill the

        planner emits to `co_lexical_alt`, which writes `Interaction.module

        = "CE"` and does *not* contribute to the CO posterior. The UX

        banner is mandatory; see `phase5_design.md §7.3`.


        `ee_alternative="speech_to_text"` plugs STT at the EE drill''s input

        layer; the rubric still scores the resulting text. `eo_alternative

        = "text_input"` swaps the EO drill for a text-input variant that

        emits `module="EE"` with a banner-disclaimed "does not measure

        real EO" caveat.


        `dyslexia_font` and `high_contrast` are UI-only signals consumed

        by the Phase 8 frontend.'
      properties:
        co_alternative:
          default: none
          enum:
          - none
          - lexical_alt
          title: Co Alternative
          type: string
        dyslexia_font:
          default: false
          title: Dyslexia Font
          type: boolean
        ee_alternative:
          default: none
          enum:
          - none
          - speech_to_text
          title: Ee Alternative
          type: string
        eo_alternative:
          default: none
          enum:
          - none
          - text_input
          title: Eo Alternative
          type: string
        high_contrast:
          default: false
          title: High Contrast
          type: boolean
      title: AccessibilityProfile
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    LoginRequest:
      additionalProperties: false
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          maxLength: 200
          minLength: 1
          title: Password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
...
    MeProfile:
      additionalProperties: false
      description: The authenticated user's full profile.
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        daily_minutes_budget:
          maximum: 480.0
          minimum: 15.0
          title: Daily Minutes Budget
          type: integer
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        email:
          format: email
          title: Email
          type: string
        id:
          format: uuid
          title: Id
          type: string
        locale:
          enum:
          - en
          - fr
          - es
          - ar
          - zh
          title: Locale
          type: string
        privacy_mode:
          description: Default 'local_only' per ADR-0017; flipping to 'cloud_optin'
            requires explicit opt-in.
          enum:
          - local_only
          - cloud_optin
          title: Privacy Mode
          type: string
        target_exam_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Target Exam Date
        target_nclc:
          maximum: 12.0
          minimum: 4.0
          title: Target Nclc
          type: integer
      required:
      - id
      - email
      - display_name
      - target_nclc
      - target_exam_date
      - daily_minutes_budget
      - locale
      - privacy_mode
      - created_at
      title: MeProfile
      type: object
...
    NCLCTrajectory:
      additionalProperties: false
      description: Time series of estimates per skill, with forecast extension.
      properties:
        horizon_end:
          format: date
          title: Horizon End
          type: string
        points:
          items:
            $ref: '#/components/schemas/NCLCTrajectoryPoint'
          title: Points
          type: array
      required:
      - horizon_end
      title: NCLCTrajectory
      type: object
    NCLCTrajectoryPoint:
      additionalProperties: false
      description: One point on the historical + forecast trajectory.
      properties:
        at:
          format: date-time
          title: At
          type: string
        ci_high:
          maximum: 12.0
          minimum: 1.0
          title: Ci High
          type: integer
        ci_low:
          maximum: 12.0
          minimum: 1.0
          title: Ci Low
          type: integer
        is_forecast:
          description: True for points that come from the planner forecast (after
            now()); False for observed history.
          title: Is Forecast
          type: boolean
        posterior_mean:
          maximum: 12.0
          minimum: 1.0
          title: Posterior Mean
          type: number
        skill:
          enum:
          - CO
          - CE
          - EE
          - EO
          title: Skill
          type: string
      required:
      - at
      - skill
      - posterior_mean
      - ci_low
      - ci_high
      - is_forecast
      title: NCLCTrajectoryPoint
      type: object
...
    RefreshRequest:
      additionalProperties: false
      properties:
        refresh_token:
          minLength: 1
          title: Refresh Token
          type: string
      required:
      - refresh_token
      title: RefreshRequest
      type: object
...
    SessionStart:
      additionalProperties: false
      description: Body of `POST /v1/session/start`.
      properties:
        drill_type:
          enum:
          - flashcard
          - cloze
          - mcq
          - shadowing
          - writing_short
          - writing_long
          - speaking_role
          - speaking_mono
          - mock_section
          - co_dictation
          - co_accent
          - co_gapfill
          - co_lexical_alt
          - ce_skim_scan
          - ce_vocab_context
          - ce_register_id
          - ce_summary
          - ee_rewrite
          - ee_connector
          - ee_error_correction
          - ee_register_adjust
          - eo_picture
          - eo_spontaneous
          - eo_roleplay
          - eo_repair
          - eo_text_alt
          title: Drill Type
          type: string
        module:
          enum:
          - CO
          - CE
          - EE
          - EO
          title: Module
          type: string
        target_minutes:
          description: Soft target; the planner may exceed for due reviews.
          maximum: 240.0
          minimum: 1.0
          title: Target Minutes
          type: integer
      required:
      - module
      - drill_type
      - target_minutes
      title: SessionStart
      type: object
    SessionState:
      additionalProperties: false
      description: The session lifecycle envelope.
      properties:
        drill_type:
          enum:
          - flashcard
          - cloze
          - mcq
          - shadowing
          - writing_short
          - writing_long
          - speaking_role
          - speaking_mono
          - mock_section
          - co_dictation
          - co_accent
          - co_gapfill
          - co_lexical_alt
          - ce_skim_scan
          - ce_vocab_context
          - ce_register_id
          - ce_summary
          - ee_rewrite
          - ee_connector
          - ee_error_correction
          - ee_register_adjust
          - eo_picture
          - eo_spontaneous
          - eo_roleplay
          - eo_repair
          - eo_text_alt
          title: Drill Type
          type: string
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        id:
          format: uuid
          title: Id
          type: string
        items_correct:
          minimum: 0.0
          title: Items Correct
          type: integer
        items_seen:
          minimum: 0.0
          title: Items Seen
          type: integer
        module:
          enum:
          - CO
          - CE
          - EE
          - EO
          title: Module
          type: string
        started_at:
          format: date-time
          title: Started At
          type: string
        target_minutes:
          maximum: 240.0
          minimum: 1.0
          title: Target Minutes
          type: integer
        user_id:
          format: uuid
          title: User Id
          type: string
      required:
      - id
      - user_id
      - module
      - drill_type
      - target_minutes
      - started_at
      - items_seen
      - items_correct
      title: SessionState
      type: object
...
    SignupRequest:
      additionalProperties: false
      properties:
        display_name:
          anyOf:
          - maxLength: 80
            type: string
          - type: 'null'
          title: Display Name
        email:
          format: email
          title: Email
          type: string
        locale:
          default: en
          pattern: ^(en|fr|es|ar|zh)$
          title: Locale
          type: string
        password:
          maxLength: 200
          minLength: 12
          title: Password
          type: string
      required:
      - email
      - password
      title: SignupRequest
      type: object
...
    TokenPair:
      additionalProperties: false
      description: JWT pair. `access_token` is short-lived; `refresh_token` is long-lived.
      properties:
        access_token:
          title: Access Token
          type: string
        access_ttl_s:
          description: Seconds until access_token expires.
          maximum: 3600.0
          minimum: 60.0
          title: Access Ttl S
          type: integer
        refresh_token:
          title: Refresh Token
          type: string
        refresh_ttl_s:
          description: Seconds until refresh_token expires.
          minimum: 3600.0
          title: Refresh Ttl S
          type: integer
        token_type:
          default: Bearer
          title: Token Type
          type: string
      required:
      - access_token
      - refresh_token
      - access_ttl_s
      - refresh_ttl_s
      title: TokenPair
      type: object
    UpdateMeRequest:
      additionalProperties: false
      description: 'Partial update of the profile. Only the supplied fields are changed.


        Sending an empty body is a no-op (returns the unchanged profile).'
      properties:
        daily_minutes_budget:
          anyOf:
          - maximum: 480.0
            minimum: 15.0
            type: integer
          - type: 'null'
          title: Daily Minutes Budget
        display_name:
          anyOf:
          - maxLength: 80
            type: string
          - type: 'null'
          title: Display Name
        locale:
          anyOf:
          - enum:
            - en
            - fr
            - es
            - ar
            - zh
            type: string
          - type: 'null'
          title: Locale
        privacy_mode:
          anyOf:
          - enum:
            - local_only
            - cloud_optin
            type: string
          - type: 'null'
          title: Privacy Mode
        target_exam_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Target Exam Date
        target_nclc:
          anyOf:
          - maximum: 12.0
            minimum: 4.0
            type: integer
          - type: 'null'
          title: Target Nclc
      title: UpdateMeRequest
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WeakPoint:
      additionalProperties: false
      description: One identified error pattern, derived from interactions + rubrics.
      properties:
        last_seen:
          format: date-time
          title: Last Seen
          type: string
        occurrences:
          minimum: 1.0
          title: Occurrences
          type: integer
        pattern:
          description: Short human-readable label, e.g. 'subjunctive after que' or
            'fr-CA accent confusable'.
          minLength: 1
          title: Pattern
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          title: Severity
          type: string
        skill:
          enum:
          - CO
          - CE
          - EE
          - EO
          title: Skill
          type: string
        suggested_drill:
          anyOf:
          - type: string
          - type: 'null'
          description: Pointer to a drill type the planner can schedule (DrillType
            enum value).
          title: Suggested Drill
      required:
      - skill
      - pattern
      - occurrences
      - last_seen
      - severity
      title: WeakPoint
      type: object
...
info:
  description: tcf-accel HTTP API. Phase 2 freezes the /v1/ contract from `02_ARCHITECTURE.md
    §2.4`; Phases 3–8 implement the handlers.
  title: tcf-accel API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/auth/login:
    post:
      description: Authenticate and return a token pair.
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '501':
          description: Phase 3 owns this implementation.
      summary: Exchange credentials for a token pair
      tags:
      - auth
  /v1/auth/refresh:
    post:
      description: Exchange a refresh token for a new access token.
      operationId: refresh_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '501':
          description: Phase 3 owns this implementation.
      summary: Refresh the access token
      tags:
      - auth
  /v1/auth/signup:
    post:
      description: Create an account and return a token pair.
      operationId: signup_v1_auth_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '501':
          description: Phase 3 owns this implementation.
      summary: Create a new account
      tags:
      - auth
...
  /v1/insights/nclc-trajectory:
    get:
      description: Return the trajectory series.
      operationId: trajectory_v1_insights_nclc_trajectory_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NCLCTrajectory'
          description: Successful Response
        '501':
          description: Phase 8 owns this implementation.
      summary: Historical per-skill NCLC + forecast
      tags:
      - insights
...
  /v1/insights/weak-points:
    get:
      description: Return the ranked weak-point list.
      operationId: weak_points_v1_insights_weak_points_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WeakPoint'
                title: Response Weak Points V1 Insights Weak Points Get
                type: array
          description: Successful Response
        '501':
          description: Phase 8 owns this implementation.
      summary: Top error patterns per skill
      tags:
      - insights
  /v1/me:
    get:
      description: Return the authenticated user's profile.
      operationId: get_me_v1_me_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeProfile'
          description: Successful Response
        '501':
          description: Phase 3 owns this implementation.
      summary: Fetch the authenticated user's profile
      tags:
      - me
    patch:
      description: Apply a partial update to the profile.
      operationId: patch_me_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeProfile'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '501':
          description: Phase 3 owns this implementation.
      summary: Partially update the authenticated user's profile
      tags:
      - me
  /v1/me/accessibility:
    get:
      description: 'Return the learner''s accessibility preferences.


        Local-only per ADR-017; the profile drives session-time drill

        routing (e.g., `co_alternative="lexical_alt"` swaps CO drills for

        the lexical alternative — `phase5_design.md §7`).'
      operationId: get_accessibility_v1_me_accessibility_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessibilityProfile'
          description: Successful Response
      summary: Fetch the learner's accessibility profile (ADR-029)
      tags:
      - me
    patch:
      description: Replace the learner's accessibility profile with the submitted
        one.
      operationId: patch_accessibility_v1_me_accessibility_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessibilityProfile'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessibilityProfile'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update the learner's accessibility profile
      tags:
      - me
...
  /v1/session/start:
    post:
      description: 'Initialize a session with the requested module and drill type.


        Enforces the exam-shape floor (ADR-028): a non-exam-shape drill is

        refused with 409 if the learner has logged zero exam-shape minutes

        in the rolling 7-day window and has not dismissed the floor this

        week.'
      operationId: start_v1_session_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionStart'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionState'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Start a practice session
      tags:
      - session
...